Building Nested Lists
Trying to just work through this on my own. Step 1 is building all the nested lists which just have all the options

step 2 will be determining how to make the content dynamic and if that can be effected in the javascript or has to be done with the python.

Might just need to pass 100 lists.